pyproject.toml
[project]
name = "imtranshelper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cnocr>=2.3.1",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
]

[build-system]
requires = ["uv_build>=0.7.9,<0.8.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "imtranshelper"
module-root = ""
